# backend/src/main/java/fr/epita/assistants/ping/api/request/UpdateUserRequest.java
package fr.epita.assistants.ping.api.request;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class UpdateUserRequest {
    private String password;
    private String displayName;
    private String avatar;
}