# backend/src/main/java/fr/epita/assistants/ping/api/response/LoginResponse.java
package fr.epita.assistants.ping.api.response;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class LoginResponse {
    private String token;
}