# backend/src/main/resources/openapi.yml
...
openapi: 3.1.0
paths:
...
  /api/user:
    post:
      summary: Create a user
      description: |
        Create a new user, its login should be unique and contain at most
        one '.' or '_'. The name should be initialized to the login split and capitalized on the separator.

        Only an admin can access this endpoint.
      tags:
      - User
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUserRequest'
        required: true
      responses:
        "200":
          description: The created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        "400":
          description: The login or the password is invalid
        "409":
          description: The login is already taken
        "403":
          description: The user is not an admin
        "401":
          description: Not Authorized
      servers:
      - url: http://localhost:8080
      security:
      - SecurityScheme:
        - admin
  /api/user/me:
    get:
      summary: Get current user
      description: |
        Retrieve the profile of the currently authenticated user using the token.
        
        The returned data corresponds to the user bound to the token.
      tags:
        - User
      responses:
        "200":
          description: The current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        "401":
          description: Not Authorized
      servers:
        - url: http://localhost:8080
      security:
        - SecurityScheme: [ ]
  /api/user/all:
    get:
      summary: List all the users
      description: |
        List all the users available, be careful the response should not contain all the
        information about the members or the owner.

        Only an admin can access this endpoint.
      tags:
      - User
      responses:
        "200":
          description: List of the users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        "403":
          description: The user is not an admin
        "401":
          description: Not Authorized
      servers:
      - url: http://localhost:8080
      security:
      - SecurityScheme:
        - admin
  /api/user/login:
    post:
      summary: Login a user
      description: |
        Login a user with its login and password.

        It should return a token if the login/password combination is valid.
      tags:
      - User
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        "200":
          description: The token of the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        "401":
          description: The login/password combination is invalid
        "400":
          description: The login or the password is null
      servers:
      - url: http://localhost:8080
  /api/user/refresh:
    get:
      summary: Refresh a token
      description: |
        Refresh the token of a user by returning a new one.

        Only a logged in user can access this endpoint.
      tags:
      - User
      responses:
        "200":
          description: The new token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        "404":
          description: The user could not be found
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      servers:
      - url: http://localhost:8080
      security:
      - SecurityScheme: []
  /api/user/{id}:
    put:
      summary: Update a user
      description: |
        Update a user

        For the sake of simplicity the password can be modified.

        Any blank value should be ignored (except for the avatar).

        Only the same user or an admin can access this endpoint.
      tags:
      - User
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
        required: true
      responses:
        "200":
          description: The modified user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        "404":
          description: The user could not be found
        "403":
          description: The user is not allowed
        "401":
          description: Not Authorized
      servers:
      - url: http://localhost:8080
      security:
      - SecurityScheme: []
    get:
      summary: Get a user
      description: |
        Get the data of a user.

        A user can only access its data while an admin can access any user data.
      tags:
      - User
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      responses:
        "200":
          description: The user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        "404":
          description: User not found
        "403":
          description: The user is not allowed to access this user
        "401":
          description: Not Authorized
      servers:
      - url: http://localhost:8080
      security:
      - SecurityScheme: []
    delete:
      summary: Delete a user
      description: |
        Delete a user.

        If the user owns projects, he cannot be deleted and a forbidden response should be returned.

        Only an admin can access this endpoint.
      tags:
      - User
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      responses:
        "204":
          description: The user was deleted
        "404":
          description: The user could not be found
        "403":
          description: "The user is not allowed to access this endpoint, or the user\
            \ owns projects"
        "401":
          description: Not Authorized
      servers:
      - url: http://localhost:8080
      security:
      - SecurityScheme:
        - admin
components:
  schemas:
...
    LoginRequest:
      type: object
      properties:
        login:
          type: string
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        token:
          type: string
...
    NewUserRequest:
      type: object
      properties:
        login:
          type: string
        password:
          type: string
        isAdmin:
          type: boolean
...
    UUID:
      type: string
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
...
    UpdateUserRequest:
      type: object
      properties:
        password:
          type: string
        displayName:
          type: string
        avatar:
          type: string
...
    UserResponse:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        login:
          type: string
        displayName:
          type: string
        isAdmin:
          type: boolean
        avatar:
          type: string
...
  securitySchemes:
    SecurityScheme:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authentication
...
info:
  title: ping API
  version: "1.0"